feat(lint): configures importas for consistent import aliases (#1078)
This linter makes sure that we use consitent aliases for imports, which are especially handy when it comes to versioned api packages. With simple rules we can ensure that e.g. "k8s.io/api/core/v1" is consistently aliased as corev1. In case of "k8s.io/apimachinery/pkg/api/errors" it will expect k8serr alias instead of apierrs, as former is more descriptive.