fix(lint): explicitly configures ireturn exceptions (#1085)
Instead of using `//nolint:ireturn` the linter allows a use of `generic`
return types (on top of the listed defaults which now have been
explicitly set).

This approach also eliminates a re-occuring issue where `nolintlint`
reports aforementioned directive as not used by defined linter.

Related issue: https://github.com/golangci/golangci-lint/issues/3228

On top of that we have `--fix` flag enabled for golangci-lint runner.
This results in removal of these comments as `nolintlint` tries to
autofix.

As a consequence the subsequent run of `make lint` yields errors for
previously disabled linters and the cycle continues :)

> [!IMPORTANT]
> This behaviour has also been observed for other linters.

As part of this the declarations of kustomize plugin constructor funcs
has been reworkted to return struct pointers instead. The reason for
this is that in our feature branch we are relying on kustomize plugins
and want to iterate over a slice of resmap.Transfomers but in the
current form these functions cannot be used due to the fact that
structs returned have pointer receivers. From the current functionality
standpoint (on incubation branch) this change is harmless.

(cherry picked from commit 613d55246d6f9b858c542d398c69f3b86c03d758)