chore(lint): skips true|false for goconst (#1110)
As we have labels and annotations with values either `true` or `false`
we often compare them in the code. This leads to complains of `goconst`
linter about extracting these "magic words" to constants. As this are
obvious values to check against there's little benefit of doing so.

This change configures `goconst` for golangci-lint runner to skip such
cases.