refactor: simplifies Feature Manifests
The raw and template manifest structs differed only in their content processing methods, making it an unnecessary abstraction. This work consolidates the two implementations, reducing duplication and resulting in cleaner code.

The implementation still adheres to the convention of using a '.tmpl.' in the file name to indicate a template and only then process it as such.

The feature builder has been reworked from chaining `ManifestLocation().Manifest` in favor of composition. This not only makes the code easier to follow but also allows to potentially mix different kinds of Manifest providers.

Additionally, `Applier` wrapper has been introduced to handle different strategies of applying manifests to the cluster based on their types.