fix: ensures SMCP is created before any other features (#1118)
With the #1052 refactoring, the order of features added to the Registry was
accidentally changed. It results in failing of metrics collection
feature which expects SMCP to be created first, but the creation runs
afterwards. The setup is eventually consistent, as the reconcile will
retry, so this not a bug per se, but results in unnecassary errors.

This fix ensures features are ordered as before and levarages
`.EnabledWhen` instead of wrapping features in `if`s.