fix: fixes authz patch injection feature precondition
If the authorization provider namespace is not specified in the DSCI the
default is constructed to be `application-namespace-auth-provider`, e.g.
`opendatahub-auth-provider`.

With the #1052 refactoring, the regression has been introduced where the value is
directly read from the spec instead of being dynamically constructed
based on the rule described above.

This is manifested with the following error, as the feature mistakenly
waits for pods across all namespaces (because of list option for
namespace being `corev1.NamespaceAll == ""`). This obviously rarely is
true, especially for large clusters.

```json
Failed applying [enable-proxy-injection-in-authorino-deployment]: 1 error occurred:
* client rate limiter Wait returned an error: context deadline exceeded
```

leading to failure of reconciling this feature.

The fix is to read the namespace from `FeatureData` instead, where the defaulting
logic is defined.

Fixes https://issues.redhat.com/browse/RHOAIENG-10268