chore: renames IngressGatewaySpec type to GatewaySpec
Original name is confusing as it does not actually define anything
related to the ingress. It is configuring Istio Gateway resource which
is, in case of Model Serving, attached to ingress service.

This change makes the change of the golang type, but not publicly
exposed `ingressGateway` config defined for KServe component.