fix(feature): preserves original target namespace (#1148)
This change ensures that original target namespace is used when invoked
directly in the builder.

Typically subsequent calls of .TargetNamespace in the feature builder
indicate coding/copy-paste error and should be reduced to one.

However,in the FeatureHandler, where we group features together we do not have
to specify target namespace for each feature, as it is defaulted to the
one defined on the handler level. This is convenient, but limits
application of features to a single namespace, as the value is always
overwritten.

(cherry picked from commit a890ffa6b32e5ec3e2c1e458adcfcab44b03e8ca)