feat(make): adds generate to default targets (#1176)
Current list of default targets include `manifests` which generates
CRDs based on API code changes, but does not invoke `generate` which is
responsible for generating additional code (namely `DeepCopy` funcs).

This can result in incomplete implemetation where CRDs reflect latest
changes in the API, but internally it might not be possible to use due
to missing copy methods or even outdated implementations of those that
already exist.

This change adds it to the list of default targets so that each code
change can be committed with complete set of changes for controller-runtime.