deploy: ApplyParams: rewrite with tempfile
Use traditional UNIX pattern of writing to a temporary file and
renaming it to the original one instead of modifying inplace.
Rename operation is atomic.

It also fixes previous code for handling unclosed files on the
filesystem (it used defer to close them on exit of the function).

Addionally add comment in case of value replaced with environment.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>