tests: e2e: do not wait deployements for not Managed components
The test retries if there are no deployements even if component is
disabled when this is a normal situation.

Make it to retry for disabled (explicitly Removed) component only if
there are deployments, but if it is not Managed instantly return
success.

It also means that the "disabled" check is inverted, it is not error
anymore.

It descreases execution time.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>