tests: e2e: testApplicationCreation cleanup (#1215)
* tests: e2e: do not print Trying again if cannot list deployments

The message is misleading, actually it returns error so will not
retry.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>

* tests: e2e: testApplicationCreation: enable immediate for PollUntilContextTimeout

After recent update the check was moved after DSC Ready check. There
is no point to wait extra Retry Interval for deployment check.

It decreases execution time.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>

* tests: e2e: remove check for dependent operator

The intention for the check originally was to return success if no
deployments for component found due to missing operator (probably
with a bug in actual check).

Anyway, the case is not valid anymore, dependent operators are
installed.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>

* tests: e2e: testApplicationCreation: early return from the loop

Simplify deployments readiness check loop with early return when
problematic deployment found and successful return as the default.

The len() == 0 check is needed otherwise it will declare success for
no deployements. It will be amended with enabled/disabled check in a
followup.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>

---------

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>