fix(crd): adds validation pattern and max length for ns fields (#1261)
This commit introduces a validation pattern and maximum length constraint
to the `Namespace` fields in our structs. We have missed it in the early
days.

Validation rules ensure that namespace names does not exceed max length
defined and are valid [RFC 1123 DNS labels](https://kubernetes.io/docs/concepts/overview/working-with-objects/namespaces/#namespaces-and-dns).