components: move logger creation to pkg/logger
Remove ConfigComponentLogger method from ComponentInterface. In the
original code this is totally internal function, not used by
external interface users.

But moreover by its functionality it prepares a logger from the
caller to be used in component's context (creates new one with
required log levels and DSC prefix or just adds the prefix to the
caller's one) so make a utility function NewNamedLogger out of
it. The actual component related configuration comes from DSC
reconciler, so wrap it as newComponentLogger there.

This is a pretty contradictive change since adding context name
idiomatically belongs to the context (component), but this allows to
do it in one place.

Alternative is to create just a final logger in the component
(l = l.WithName(componentName)).

This patch changes dashboard's log name in downstream.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>