DSCI CRD: allow default as a value for logmode
default is a separate mode in the sources, not just selection of one
of the others, so allow it explicitly.

There is no difference if it is set to "" explicitly or not set
since "" is a string default value.

In the current code as soon as devFlags are present, the mode is
taken in use, so it is reset to default even if it is not set which
does not look right (I would expect leave it untouched if is not
set). If it is fixed, then it will be impossible to reset logging
mode to default from devel or prod: if "" is considered as "not set
explicitly", then there is no value for default, the object checker
rejects values not in list.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>