chore: use same version of golangci-lint as GHA
The GitHub action has updated past the version of golangci-lint that
we've specified in the Makefile. This brings them to the same version
to prevent inconsistency (moving to Go 1.22 will introduce such
inconsistency).

The task to install this tool currently doesn't ensure the version, if
it's already in place, so one can to run `make -B golangci-lint` once
to get the correct versions.