Add custom CRD deployment logic (#1418)
* Add custom CRD deployment logic

The CRDs are meant to be left in the cluster, so they are not tight to
components' lifecycle hence, CRDs requires some custom handling so the
metadata is not cluttered by useless component/platform annotations and
labels.

As a conseguence, the each component has to set-up a dedicated handler
and predicate in order to mape and filter out relevant CRDs events.

* Fix tests

* Fix tests

* Fix findings