Specify full x.y.z version of Go in go.mod (#1442)
JIRA: https://issues.redhat.com/browse/RHOAIENG-16819

It seems like on some platforms, when the local installation
installation of Go decides that it needs to download a different
toolchain, that it may fail unless we either specify a full x.y.z
version in the `go` directive, or explicitly specify the `toolchain`
directive.

I believe for our uses, using a fully specified x.y.z `go` directive
value is preferable, to prevent any issues with explicitly defining a
toolchain where we might have a newer version in a container image
build that has a default of `GOTOOLCHAIN='local'` (I'm not 100% sure
if that would cause any issues, but I'm 100% sure that the way
proposed in this commit shouldn't).