Update golang.org/x/net to v0.33.0 (#1463)
This change is to update to a version that isn't flagged as affected
by CVE-2024-45338.

While we know that this operator doesn't call the vulnerable code
paths (hooray for `govulncheck`), some scanning tools are not yet
smart enough to be able to discern that. So it's preferable just to
update.

Go advisory: https://pkg.go.dev/vuln/GO-2024-3333