Show KServe defaultDeploymentMode in DSC status
JIRA: https://issues.redhat.com/browse/RHOAIENG-16240

The dashboard would like to be able to read this value, and should
only read it from the status.

The value for the status is always read from the KServe ConfigMap,
rather than ever assuming that what is set in the DSC, or what the
operator defaults to, is correct: this is because that ConfigMap can
be annotated to be unmanaged by the operator, in which case a user may
have changed the value to something else, and the status should
reflect this reality.