get_all_manifests.sh: use key (component name) as the target folder (#1456)
There is no point to set target folder explicitly if it can be the
same as the component name.

The patch is a step to unification of component names.

In the error message for overriding with an option also
`branch-name` replaced to `ref-name`. Leftover.

For odh-notebook-controller subcomponents directory is a part of the
key.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>