Simplify DataScienceCluster conditions
This commit aims to simpplify the DataScienceCluster status conditions
by removing transient conditions (Reconciling, Deploying, etc) and
replacing them with two conditsion:

- Available that is used to reflects the work done by the controller, so
  an Available condition with status False would signal that the
  controller encountered some troubles while reconciling the status of
  the cluster
- Ready that is used to reflect the status of the component managed by
  the controller.

It is perfectly possible to be in a state where Available=False and
Ready=True which can happen in case the controller has troubles to
perform some reconciliation logic because of i.e. a transient kubernetes
error, but at the same time all the components are running.

The object would then result in soimething like:

    apiVersion: datasciencecluster.opendatahub.io/v1
    kind: DataScienceCluster
    metadata:
        name: default-dsc
    spec: {}
    status:
        conditions:
        - lastHeartbeatTime: "2025-01-15T12:54:20Z"
            lastTransitionTime: "2025-01-15T12:54:20Z"
            message: DataScienceCluster resource reconciled successfully
            reason: Available
            status: "True"
            type: Available
        - lastHeartbeatTime: "2025-01-15T12:54:20Z"
            lastTransitionTime: "2025-01-15T12:54:20Z"
            message: Ready
            reason: Ready
            status: "True"
            type: Ready

It is possible to wait for the DataScienceCluster to be ready by using
the kubectl wait command:

    kubectl wait --for=condition=ready datascienceclusters.datasciencecluster.opendatahub.io default-dsc