componentsregistry: use structure to store handlers (#1502)
To make it possible to mock registry in DSC reconcileComponents for
testing, implement registry as a structure. Make package level
wrappers on Add and ForEach methods which operate on the default
statically created registry.

This allows keep working with the default registry in a real system
without extra arguments like before. It's in init()s and in
main(). But code under test can accept testing version when needed.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>