Improve e2e test configurability (#1650)
* Improve e2e test configurability

This feature is to improve the configurability of the e2e test suite,
which is particularly useful when running the test suite locally.

There are two new flags to globally enabled/disable some test grouops
such as components and services:

* --test-components=true|false
* --test-services=true|false

It is also possible to disable a particular test by preceding its name
with an exlamation mark (!), i.e. to run only the test for the dashbaord
component and skip the tests for the auth service:

  make e2e-test E2E_TEST_FLAGS='--test-component=dashboard --test-service=!auth'

* Fix validation logic