chore: Allow cross-compilation in Dockerfile (#1681)
My machine runs on aarch64, but I want to be able to run container
images that I build during development on x86_64 clusters.

The main problem is that the default is to use the TARGETPLATFORM for
all stages in an image build, which means that, for example, the `go`
commands that are run in the `builder` stage will be using an x86_64
`go` binary, which essentially doesn't work on aarch64.

When this instead specifies to use BUILDPLATFORM for those stages
before the last one, it defaults to using the local platform (os/arch)
for those containers, so I get aarch64 builder stage container images,
containing aarch64 `go` binaries.

Since I specify `--platform=linux/amd64` when doing the `podman
build`, it will still build an aarch64 binary of the operator, and put
that into an aarch64 final image.

The full command that I use to build the image is as follows.

```commandline
podman build --platform linux/amd64 -f Dockerfiles/Dockerfile \
  --build-arg CGO_ENABLED=0 \
  -t odh-operator:latest .
```

Notice that I also have to specify CGO_ENABLED=0, otherwise it'll
complain with the following error at the `go build` command:

> gcc: error: unrecognized command line option '-m64'