Add tests around FTs in Kserve before removal (#1697)
JIRA: https://issues.redhat.com/browse/RHOAIENG-18045

Some of these tests will change when the code changes, and the hope is
that those test changes will give more clarity and confidence in the
code changes.

The commented entries in these tests are because the authorino
operator is not deployed in the e2e tests in OpenShift CI, so those
resources are not deployed.