[sync] Remove FeatureTracker code from Kserve reconciler (#1772)
* fix(test): ensure jq matchers can handle null values (#1698)

* Add tests around FTs in Kserve before removal (#1697)

JIRA: https://issues.redhat.com/browse/RHOAIENG-18045

Some of these tests will change when the code changes, and the hope is
that those test changes will give more clarity and confidence in the
code changes.

The commented entries in these tests are because the authorino
operator is not deployed in the e2e tests in OpenShift CI, so those
resources are not deployed.

* Refactor the gc action/service (#1747)

The original GC action/service was designed under the assumption that
all Kubernetes resource types the operator manages were known at
startup and remained unchanged over time. As a result, a single GC
service was shared across multiple GC actions.

However, with the removal of feature trackers and the need of dynamic
dependencies (e.g., Serverless, Service Mesh), this static resource
list assumption is no longer valid.

To address this, the shared GC service has been removed, and the GC
engine is now integrated directly within each GC action.

* Remove FeatureTracker code from Kserve reconciler (#1521)

* Add FeatureTracker as a legacy ownerref kind

JIRA: https://issues.redhat.com/browse/RHOAIENG-18045

This will cause the deploy action to replace any existing
FeatureTracker owner ref on a resource with the component as the owner
instead.

* Remove FeatureTracker code from Kserve reconciler

Jira: https://issues.redhat.com/browse/RHOAIENG-18045

* Deduplicate KnativeServing template in Kserve

* Add dependency label to Kserve templates

* Always render all kserve templates, process after

There are some intricacies around how transitions need to be handled
for the resources that get conditionally created for
servicemesh/serverless.

- managed -> unmanaged:
  In this case, the resources should NOT be deleted, just not
  reconciled any longer. So to prevent the GC from collecting them,
  the ownerrefs and part-of labels are removed.
  Additionally the resources should be removed from the rr.Resources
  so that they are no longer reconciled.
- the resources should be removed from the rr.Resources
  so that they are not reconciled and can be garbage collected.

* Add tests around kserve serving managementState

* Use rr.RemoveResources rather than slices.DeleteFunc

* Add dummy existing FTs at start of Kserve e2e test

---------

Co-authored-by: Luca Burgazzoli <lburgazzoli@users.noreply.github.com>